fix uint8_t conversion to string

+++ src/main.cpp
@@ -46,12 +46,12 @@ int main(int argc, char *argv[])
 	// setup inputs and outputs
 	for(auto pin : config.inputs) {
 		pinMode(pin, INPUT);
-		std::cout << "pin " << pin << " set as INPUT" << std::endl;
+		std::cout << "pin " << static_cast<int>(pin) << " set as INPUT" << std::endl;
 	}
 
 	for(auto pin : config.outputs) {
 		pinMode(pin, OUTPUT);
-		std::cout << "pin " << pin << " set as OUTPUT" << std::endl;
+		std::cout << "pin " << static_cast<int>(pin) << " set as OUTPUT" << std::endl;
 	}
 
 	GPIOCommandHandler commandHandler(config.outputs);
